Fix flaky test_debug_info.
Pull Request resolved: https://github.com/pytorch/pytorch/pull/31675

This test could be flaky since there could be inflight RPC requests as
part of startup which might not have finished. As a result, if they finish
between the different calls to retrieve debug_info, there could be a problem
since we would report separate information. As a result, we wait to ensure
the metrics stabilize to avoid flakiness.
ghstack-source-id: 96188488

Differential Revision: [D19242588](https://our.internmc.facebook.com/intern/diff/D19242588/)